Returns a list with data.frames with stature estimations.

Returns a list with data.frames of stature estimations for each specified formula and the data provided in a data.frame of class statuaar_data_table.

Usage

getStature(shortnames, statuaar_data_table)

Arguments

shortnames

A vector of short names of the formula, e.g. c('bb65', 'tg01'), "all" for all formulas.

statuaar_data_table

A data.frame of class statuaar_data_table as provided by prep.statuaar.data(). #'

Ind: identifyer for each individual.

Sex: sex of the individual. Accepted values: 1 (male), 2 (female), 3 (indet) or m(ale), f(emale), indet.

variable: short name of the measure

value: measured value.

Value

A list with a data.frame for each specified stature estimation formula. The item name corresponds to the specified formula. Each data.frame with seven columns:

ind: individual identifyer (rownames),

sex: as provided for calculation: m, f, indet.

stature: estimated on the provided sex and bone measures,

bone measure name(s): bones used for calculation,

female stature: columns with alternative stature for three sex classes,

male stature,

indet. stature and

n_measures: number of bone measures included: e.g. 2 Fem2 (left, right) + 1 Tib1

Author

Christoph Rinne crinne@ufg.uni-kiel.de

Examples

# Get the bone measures needed for the functions.
getStature(c('bb65', 'tg01'), dl.trotter.gleser)
#> $bb65
#>      sex stature       bone if_female if_male if_indet n_measures
#> 1_1    m    1551 Hum1, Fem1      1557    1551     1554          2
#> 1_10   m    1617 Hum1, Fem1      1609    1617     1613          2
#> 1_11   m    1629 Hum1, Fem1      1618    1629     1623          2
#> 1_12   m    1637 Hum1, Fem1      1625    1637     1631          2
#> 1_13   m    1645 Hum1, Fem1      1631    1645     1638          2
#> 1_14   m    1652 Hum1, Fem1      1637    1652     1644          2
#> 1_15   m    1660 Hum1, Fem1      1643    1660     1652          2
#> 1_16   m    1669 Hum1, Fem1      1650    1669     1659          2
#> 1_17   m    1676 Hum1, Fem1      1656    1676     1666          2
#> 1_18   m    1684 Hum1, Fem1      1661    1684     1673          2
#> 1_19   m    1691 Hum1, Fem1      1667    1691     1679          2
#> 1_2    m    1558 Hum1, Fem1      1562    1558     1560          2
#> 1_20   m    1700 Hum1, Fem1      1675    1700     1688          2
#> 1_21   m    1708 Hum1, Fem1      1681    1708     1694          2
#> 1_22   m    1715 Hum1, Fem1      1686    1715     1701          2
#> 1_23   m    1723 Hum1, Fem1      1692    1723     1707          2
#> 1_24   m    1730 Hum1, Fem1      1698    1730     1714          2
#> 1_25   m    1739 Hum1, Fem1      1706    1739     1722          2
#> 1_26   m    1747 Hum1, Fem1      1711    1747     1729          2
#> 1_27   m    1754 Hum1, Fem1      1717    1754     1736          2
#> 1_28   m    1761 Hum1, Fem1      1723    1761     1742          2
#> 1_29   m    1770 Hum1, Fem1      1730    1770     1750          2
#> 1_3    m    1565 Hum1, Fem1      1568    1565     1567          2
#> 1_30   m    1778 Hum1, Fem1      1736    1778     1757          2
#> 1_31   m    1786 Hum1, Fem1      1742    1786     1764          2
#> 1_32   m    1793 Hum1, Fem1      1748    1793     1770          2
#> 1_33   m    1802 Hum1, Fem1      1755    1802     1778          2
#> 1_34   m    1809 Hum1, Fem1      1761    1809     1785          2
#> 1_35   m    1817 Hum1, Fem1      1767    1817     1792          2
#> 1_36   m    1825 Hum1, Fem1      1773    1825     1799          2
#> 1_37   m    1833 Hum1, Fem1      1780    1833     1807          2
#> 1_38   m    1841 Hum1, Fem1      1786    1841     1813          2
#> 1_39   m    1848 Hum1, Fem1      1791    1848     1820          2
#> 1_4    m    1574 Hum1, Fem1      1575    1574     1575          2
#> 1_40   m    1856 Hum1, Fem1      1798    1856     1827          2
#> 1_41   m    1865 Hum1, Fem1      1805    1865     1835          2
#> 1_42   m    1872 Hum1, Fem1      1810    1872     1841          2
#> 1_43   m    1880 Hum1, Fem1      1816    1880     1848          2
#> 1_44   m    1887 Hum1, Fem1      1822    1887     1855          2
#> 1_45   m    1897 Hum1, Fem1      1830    1897     1863          2
#> 1_46   m    1904 Hum1, Fem1      1835    1904     1870          2
#> 1_47   m    1911 Hum1, Fem1      1841    1911     1876          2
#> 1_5    m    1582 Hum1, Fem1      1581    1582     1582          2
#> 1_6    m    1590 Hum1, Fem1      1587    1590     1588          2
#> 1_7    m    1597 Hum1, Fem1      1593    1597     1595          2
#> 1_8    m    1606 Hum1, Fem1      1600    1606     1603          2
#> 1_9    m    1613 Hum1, Fem1      1606    1613     1609          2
#> 2_1    m    1570 Hum1, Fem1      1572    1570     1571          2
#> 2_10   m    1642 Hum1, Fem1      1629    1642     1636          2
#> 2_11   m    1650 Hum1, Fem1      1635    1650     1642          2
#> 2_12   m    1659 Hum1, Fem1      1642    1659     1651          2
#> 2_13   m    1668 Hum1, Fem1      1649    1668     1658          2
#> 2_14   m    1676 Hum1, Fem1      1655    1676     1665          2
#> 2_15   m    1683 Hum1, Fem1      1661    1683     1672          2
#> 2_16   m    1691 Hum1, Fem1      1668    1691     1679          2
#> 2_17   m    1699 Hum1, Fem1      1674    1699     1687          2
#> 2_18   m    1708 Hum1, Fem1      1680    1708     1694          2
#> 2_19   m    1715 Hum1, Fem1      1686    1715     1701          2
#> 2_2    m    1578 Hum1, Fem1      1578    1578     1578          2
#> 2_20   m    1723 Hum1, Fem1      1693    1723     1708          2
#> 2_21   m    1731 Hum1, Fem1      1699    1731     1715          2
#> 2_22   m    1739 Hum1, Fem1      1706    1739     1723          2
#> 2_23   m    1747 Hum1, Fem1      1711    1747     1729          2
#> 2_24   m    1755 Hum1, Fem1      1718    1755     1736          2
#> 2_25   m    1763 Hum1, Fem1      1724    1763     1744          2
#> 2_26   m    1771 Hum1, Fem1      1731    1771     1751          2
#> 2_27   m    1780 Hum1, Fem1      1738    1780     1759          2
#> 2_28   m    1788 Hum1, Fem1      1744    1788     1766          2
#> 2_29   m    1796 Hum1, Fem1      1751    1796     1774          2
#> 2_3    m    1586 Hum1, Fem1      1584    1586     1585          2
#> 2_30   m    1805 Hum1, Fem1      1757    1805     1781          2
#> 2_31   m    1812 Hum1, Fem1      1763    1812     1787          2
#> 2_32   m    1820 Hum1, Fem1      1769    1820     1795          2
#> 2_33   m    1828 Hum1, Fem1      1776    1828     1802          2
#> 2_34   m    1836 Hum1, Fem1      1782    1836     1809          2
#> 2_35   m    1844 Hum1, Fem1      1788    1844     1816          2
#> 2_36   m    1852 Hum1, Fem1      1795    1852     1823          2
#> 2_37   m    1860 Hum1, Fem1      1801    1860     1831          2
#> 2_38   m    1868 Hum1, Fem1      1807    1868     1838          2
#> 2_39   m    1876 Hum1, Fem1      1813    1876     1845          2
#> 2_4    m    1594 Hum1, Fem1      1591    1594     1593          2
#> 2_40   m    1884 Hum1, Fem1      1820    1884     1852          2
#> 2_41   m    1892 Hum1, Fem1      1826    1892     1859          2
#> 2_42   m    1900 Hum1, Fem1      1833    1900     1866          2
#> 2_43   m    1909 Hum1, Fem1      1840    1909     1874          2
#> 2_44   m    1917 Hum1, Fem1      1846    1917     1882          2
#> 2_45   m    1925 Hum1, Fem1      1852    1925     1889          2
#> 2_46   m    1933 Hum1, Fem1      1859    1933     1896          2
#> 2_47   m    1941 Hum1, Fem1      1865    1941     1903          2
#> 2_5    m    1602 Hum1, Fem1      1597    1602     1600          2
#> 2_6    m    1610 Hum1, Fem1      1603    1610     1606          2
#> 2_7    m    1618 Hum1, Fem1      1610    1618     1614          2
#> 2_8    m    1626 Hum1, Fem1      1616    1626     1621          2
#> 2_9    m    1634 Hum1, Fem1      1623    1634     1628          2
#> 3_1    f    1513 Hum1, Fem1      1513    1495     1504          2
#> 3_10   f    1565 Hum1, Fem1      1565    1561     1563          2
#> 3_11   f    1571 Hum1, Fem1      1571    1569     1570          2
#> 3_12   f    1576 Hum1, Fem1      1576    1576     1576          2
#> 3_13   f    1582 Hum1, Fem1      1582    1583     1583          2
#> 3_14   f    1588 Hum1, Fem1      1588    1591     1589          2
#> 3_15   f    1594 Hum1, Fem1      1594    1598     1596          2
#> 3_16   f    1600 Hum1, Fem1      1600    1606     1603          2
#> 3_17   f    1606 Hum1, Fem1      1606    1614     1610          2
#> 3_18   f    1612 Hum1, Fem1      1612    1621     1616          2
#> 3_19   f    1618 Hum1, Fem1      1618    1628     1623          2
#> 3_2    f    1518 Hum1, Fem1      1518    1502     1510          2
#> 3_20   f    1624 Hum1, Fem1      1624    1636     1630          2
#> 3_21   f    1629 Hum1, Fem1      1629    1643     1636          2
#> 3_22   f    1635 Hum1, Fem1      1635    1650     1643          2
#> 3_23   f    1641 Hum1, Fem1      1641    1658     1649          2
#> 3_24   f    1647 Hum1, Fem1      1647    1665     1656          2
#> 3_25   f    1653 Hum1, Fem1      1653    1672     1663          2
#> 3_26   f    1658 Hum1, Fem1      1658    1680     1669          2
#> 3_27   f    1664 Hum1, Fem1      1664    1687     1676          2
#> 3_28   f    1669 Hum1, Fem1      1669    1693     1681          2
#> 3_29   f    1675 Hum1, Fem1      1675    1700     1688          2
#> 3_3    f    1524 Hum1, Fem1      1524    1510     1517          2
#> 3_30   f    1681 Hum1, Fem1      1681    1708     1694          2
#> 3_31   f    1686 Hum1, Fem1      1686    1715     1701          2
#> 3_32   f    1692 Hum1, Fem1      1692    1723     1707          2
#> 3_33   f    1698 Hum1, Fem1      1698    1730     1714          2
#> 3_34   f    1704 Hum1, Fem1      1704    1737     1721          2
#> 3_35   f    1710 Hum1, Fem1      1710    1745     1727          2
#> 3_36   f    1715 Hum1, Fem1      1715    1752     1734          2
#> 3_37   f    1722 Hum1, Fem1      1722    1760     1741          2
#> 3_38   f    1728 Hum1, Fem1      1728    1767     1748          2
#> 3_39   f    1734 Hum1, Fem1      1734    1775     1754          2
#> 3_4    f    1530 Hum1, Fem1      1530    1517     1524          2
#> 3_40   f    1739 Hum1, Fem1      1739    1782     1761          2
#> 3_41   f    1745 Hum1, Fem1      1745    1790     1767          2
#> 3_42   f    1751 Hum1, Fem1      1751    1797     1774          2
#> 3_43   f    1757 Hum1, Fem1      1757    1804     1780          2
#> 3_44   f    1763 Hum1, Fem1      1763    1812     1787          2
#> 3_45   f    1768 Hum1, Fem1      1768    1819     1794          2
#> 3_5    f    1536 Hum1, Fem1      1536    1524     1530          2
#> 3_6    f    1542 Hum1, Fem1      1542    1532     1537          2
#> 3_7    f    1547 Hum1, Fem1      1547    1539     1543          2
#> 3_8    f    1553 Hum1, Fem1      1553    1547     1550          2
#> 3_9    f    1559 Hum1, Fem1      1559    1554     1556          2
#> 4_1    f    1516 Hum1, Fem1      1516    1500     1508          2
#> 4_10   f    1573 Hum1, Fem1      1573    1571     1572          2
#> 4_11   f    1579 Hum1, Fem1      1579    1579     1579          2
#> 4_12   f    1585 Hum1, Fem1      1585    1587     1586          2
#> 4_13   f    1592 Hum1, Fem1      1592    1596     1594          2
#> 4_14   f    1598 Hum1, Fem1      1598    1603     1601          2
#> 4_15   f    1604 Hum1, Fem1      1604    1611     1607          2
#> 4_16   f    1610 Hum1, Fem1      1610    1619     1615          2
#> 4_17   f    1617 Hum1, Fem1      1617    1628     1623          2
#> 4_18   f    1623 Hum1, Fem1      1623    1635     1629          2
#> 4_19   f    1630 Hum1, Fem1      1630    1643     1636          2
#> 4_2    f    1522 Hum1, Fem1      1522    1507     1515          2
#> 4_20   f    1635 Hum1, Fem1      1635    1651     1643          2
#> 4_21   f    1643 Hum1, Fem1      1643    1660     1652          2
#> 4_22   f    1649 Hum1, Fem1      1649    1668     1658          2
#> 4_23   f    1655 Hum1, Fem1      1655    1675     1665          2
#> 4_24   f    1661 Hum1, Fem1      1661    1683     1672          2
#> 4_25   f    1667 Hum1, Fem1      1667    1690     1679          2
#> 4_26   f    1674 Hum1, Fem1      1674    1700     1687          2
#> 4_27   f    1680 Hum1, Fem1      1680    1707     1694          2
#> 4_28   f    1686 Hum1, Fem1      1686    1715     1700          2
#> 4_29   f    1692 Hum1, Fem1      1692    1723     1708          2
#> 4_3    f    1529 Hum1, Fem1      1529    1515     1522          2
#> 4_30   f    1699 Hum1, Fem1      1699    1732     1715          2
#> 4_31   f    1706 Hum1, Fem1      1706    1740     1723          2
#> 4_32   f    1712 Hum1, Fem1      1712    1747     1729          2
#> 4_33   f    1717 Hum1, Fem1      1717    1754     1736          2
#> 4_34   f    1725 Hum1, Fem1      1725    1764     1744          2
#> 4_35   f    1731 Hum1, Fem1      1731    1771     1751          2
#> 4_36   f    1737 Hum1, Fem1      1737    1779     1758          2
#> 4_37   f    1743 Hum1, Fem1      1743    1787     1765          2
#> 4_38   f    1750 Hum1, Fem1      1750    1796     1773          2
#> 4_39   f    1756 Hum1, Fem1      1756    1804     1780          2
#> 4_4    f    1534 Hum1, Fem1      1534    1523     1528          2
#> 4_40   f    1762 Hum1, Fem1      1762    1811     1787          2
#> 4_41   f    1768 Hum1, Fem1      1768    1818     1793          2
#> 4_42   f    1775 Hum1, Fem1      1775    1828     1802          2
#> 4_43   f    1781 Hum1, Fem1      1781    1835     1808          2
#> 4_44   f    1788 Hum1, Fem1      1788    1843     1816          2
#> 4_45   f    1794 Hum1, Fem1      1794    1851     1822          2
#> 4_5    f    1541 Hum1, Fem1      1541    1531     1536          2
#> 4_6    f    1548 Hum1, Fem1      1548    1539     1544          2
#> 4_7    f    1553 Hum1, Fem1      1553    1547     1550          2
#> 4_8    f    1560 Hum1, Fem1      1560    1555     1557          2
#> 4_9    f    1567 Hum1, Fem1      1567    1564     1565          2
#> 
#> $tg01
#>      sex stature    bone if_female if_male if_indet n_measures
#> 1_1    m    1507 4. Fem1      1466    1507     1487          1
#> 1_10   m    1588 4. Fem1      1553    1588     1570          1
#> 1_11   m    1596 4. Fem1      1562    1596     1579          1
#> 1_12   m    1604 4. Fem1      1571    1604     1588          1
#> 1_13   m    1613 4. Fem1      1580    1613     1597          1
#> 1_14   m    1621 4. Fem1      1589    1621     1605          1
#> 1_15   m    1632 4. Fem1      1601    1632     1616          1
#> 1_16   m    1640 4. Fem1      1610    1640     1625          1
#> 1_17   m    1649 4. Fem1      1619    1649     1634          1
#> 1_18   m    1657 4. Fem1      1628    1657     1643          1
#> 1_19   m    1666 4. Fem1      1637    1666     1651          1
#> 1_2    m    1516 4. Fem1      1475    1516     1496          1
#> 1_20   m    1676 4. Fem1      1649    1676     1662          1
#> 1_21   m    1685 4. Fem1      1658    1685     1671          1
#> 1_22   m    1693 4. Fem1      1667    1693     1680          1
#> 1_23   m    1702 4. Fem1      1676    1702     1689          1
#> 1_24   m    1710 4. Fem1      1685    1710     1698          1
#> 1_25   m    1721 4. Fem1      1697    1721     1709          1
#> 1_26   m    1729 4. Fem1      1706    1729     1717          1
#> 1_27   m    1737 4. Fem1      1715    1737     1726          1
#> 1_28   m    1746 4. Fem1      1724    1746     1735          1
#> 1_29   m    1754 4. Fem1      1733    1754     1744          1
#> 1_3    m    1524 4. Fem1      1485    1524     1504          1
#> 1_30   m    1765 4. Fem1      1744    1765     1755          1
#> 1_31   m    1773 4. Fem1      1754    1773     1763          1
#> 1_32   m    1782 4. Fem1      1763    1782     1772          1
#> 1_33   m    1790 4. Fem1      1772    1790     1781          1
#> 1_34   m    1799 4. Fem1      1781    1799     1790          1
#> 1_35   m    1809 4. Fem1      1792    1809     1801          1
#> 1_36   m    1818 4. Fem1      1801    1818     1810          1
#> 1_37   m    1826 4. Fem1      1811    1826     1818          1
#> 1_38   m    1834 4. Fem1      1820    1834     1827          1
#> 1_39   m    1843 4. Fem1      1829    1843     1836          1
#> 1_4    m    1533 4. Fem1      1494    1533     1513          1
#> 1_40   m    1853 4. Fem1      1840    1853     1847          1
#> 1_41   m    1862 4. Fem1      1849    1862     1856          1
#> 1_42   m    1870 4. Fem1      1858    1870     1864          1
#> 1_43   m    1879 4. Fem1      1868    1879     1873          1
#> 1_44   m    1887 4. Fem1      1877    1887     1882          1
#> 1_45   m    1898 4. Fem1      1888    1898     1893          1
#> 1_46   m    1906 4. Fem1      1897    1906     1902          1
#> 1_47   m    1915 4. Fem1      1906    1915     1910          1
#> 1_5    m    1543 4. Fem1      1505    1543     1524          1
#> 1_6    m    1552 4. Fem1      1514    1552     1533          1
#> 1_7    m    1560 4. Fem1      1523    1560     1542          1
#> 1_8    m    1569 4. Fem1      1532    1569     1550          1
#> 1_9    m    1577 4. Fem1      1542    1577     1559          1
#> 2_1    m    1520 4. Fem1      1480    1520     1500          1
#> 2_10   m    1611 4. Fem1      1578    1611     1594          1
#> 2_11   m    1619 4. Fem1      1587    1619     1603          1
#> 2_12   m    1630 4. Fem1      1599    1630     1614          1
#> 2_13   m    1640 4. Fem1      1610    1640     1625          1
#> 2_14   m    1651 4. Fem1      1621    1651     1636          1
#> 2_15   m    1659 4. Fem1      1630    1659     1645          1
#> 2_16   m    1670 4. Fem1      1642    1670     1656          1
#> 2_17   m    1680 4. Fem1      1653    1680     1667          1
#> 2_18   m    1691 4. Fem1      1665    1691     1678          1
#> 2_19   m    1699 4. Fem1      1674    1699     1687          1
#> 2_2    m    1529 4. Fem1      1489    1529     1509          1
#> 2_20   m    1710 4. Fem1      1685    1710     1698          1
#> 2_21   m    1721 4. Fem1      1697    1721     1709          1
#> 2_22   m    1731 4. Fem1      1708    1731     1720          1
#> 2_23   m    1740 4. Fem1      1717    1740     1728          1
#> 2_24   m    1750 4. Fem1      1728    1750     1739          1
#> 2_25   m    1761 4. Fem1      1740    1761     1750          1
#> 2_26   m    1771 4. Fem1      1751    1771     1761          1
#> 2_27   m    1780 4. Fem1      1760    1780     1770          1
#> 2_28   m    1790 4. Fem1      1772    1790     1781          1
#> 2_29   m    1801 4. Fem1      1783    1801     1792          1
#> 2_3    m    1539 4. Fem1      1500    1539     1520          1
#> 2_30   m    1811 4. Fem1      1795    1811     1803          1
#> 2_31   m    1820 4. Fem1      1804    1820     1812          1
#> 2_32   m    1830 4. Fem1      1815    1830     1823          1
#> 2_33   m    1841 4. Fem1      1827    1841     1834          1
#> 2_34   m    1851 4. Fem1      1838    1851     1845          1
#> 2_35   m    1860 4. Fem1      1847    1860     1853          1
#> 2_36   m    1870 4. Fem1      1858    1870     1864          1
#> 2_37   m    1881 4. Fem1      1870    1881     1875          1
#> 2_38   m    1891 4. Fem1      1881    1891     1886          1
#> 2_39   m    1900 4. Fem1      1890    1900     1895          1
#> 2_4    m    1550 4. Fem1      1512    1550     1531          1
#> 2_40   m    1910 4. Fem1      1902    1910     1906          1
#> 2_41   m    1921 4. Fem1      1913    1921     1917          1
#> 2_42   m    1932 4. Fem1      1925    1932     1928          1
#> 2_43   m    1940 4. Fem1      1934    1940     1937          1
#> 2_44   m    1951 4. Fem1      1945    1951     1948          1
#> 2_45   m    1961 4. Fem1      1956    1961     1959          1
#> 2_46   m    1972 4. Fem1      1968    1972     1970          1
#> 2_47   m    1980 4. Fem1      1977    1980     1979          1
#> 2_5    m    1560 4. Fem1      1523    1560     1542          1
#> 2_6    m    1569 4. Fem1      1532    1569     1550          1
#> 2_7    m    1579 4. Fem1      1544    1579     1561          1
#> 2_8    m    1590 4. Fem1      1555    1590     1572          1
#> 2_9    m    1600 4. Fem1      1567    1600     1583          1
#> 3_1    f    1391 4. Fem1      1391    1438     1414          1
#> 3_10   f    1473 4. Fem1      1473    1514     1493          1
#> 3_11   f    1482 4. Fem1      1482    1522     1502          1
#> 3_12   f    1491 4. Fem1      1491    1531     1511          1
#> 3_13   f    1500 4. Fem1      1500    1539     1520          1
#> 3_14   f    1510 4. Fem1      1510    1548     1529          1
#> 3_15   f    1519 4. Fem1      1519    1556     1537          1
#> 3_16   f    1530 4. Fem1      1530    1566     1548          1
#> 3_17   f    1539 4. Fem1      1539    1575     1557          1
#> 3_18   f    1548 4. Fem1      1548    1583     1566          1
#> 3_19   f    1557 4. Fem1      1557    1592     1575          1
#> 3_2    f    1400 4. Fem1      1400    1446     1423          1
#> 3_20   f    1567 4. Fem1      1567    1600     1583          1
#> 3_21   f    1576 4. Fem1      1576    1609     1592          1
#> 3_22   f    1585 4. Fem1      1585    1617     1601          1
#> 3_23   f    1594 4. Fem1      1594    1626     1610          1
#> 3_24   f    1603 4. Fem1      1603    1634     1619          1
#> 3_25   f    1612 4. Fem1      1612    1642     1627          1
#> 3_26   f    1621 4. Fem1      1621    1651     1636          1
#> 3_27   f    1630 4. Fem1      1630    1659     1645          1
#> 3_28   f    1640 4. Fem1      1640    1668     1654          1
#> 3_29   f    1649 4. Fem1      1649    1676     1662          1
#> 3_3    f    1409 4. Fem1      1409    1455     1432          1
#> 3_30   f    1658 4. Fem1      1658    1685     1671          1
#> 3_31   f    1667 4. Fem1      1667    1693     1680          1
#> 3_32   f    1676 4. Fem1      1676    1702     1689          1
#> 3_33   f    1685 4. Fem1      1685    1710     1698          1
#> 3_34   f    1694 4. Fem1      1694    1718     1706          1
#> 3_35   f    1703 4. Fem1      1703    1727     1715          1
#> 3_36   f    1713 4. Fem1      1713    1735     1724          1
#> 3_37   f    1724 4. Fem1      1724    1746     1735          1
#> 3_38   f    1733 4. Fem1      1733    1754     1744          1
#> 3_39   f    1742 4. Fem1      1742    1763     1752          1
#> 3_4    f    1418 4. Fem1      1418    1463     1441          1
#> 3_40   f    1751 4. Fem1      1751    1771     1761          1
#> 3_41   f    1760 4. Fem1      1760    1780     1770          1
#> 3_42   f    1770 4. Fem1      1770    1788     1779          1
#> 3_43   f    1779 4. Fem1      1779    1796     1788          1
#> 3_44   f    1788 4. Fem1      1788    1805     1796          1
#> 3_45   f    1797 4. Fem1      1797    1813     1805          1
#> 3_5    f    1428 4. Fem1      1428    1472     1450          1
#> 3_6    f    1437 4. Fem1      1437    1480     1458          1
#> 3_7    f    1446 4. Fem1      1446    1488     1467          1
#> 3_8    f    1455 4. Fem1      1455    1497     1476          1
#> 3_9    f    1464 4. Fem1      1464    1505     1485          1
#> 4_1    f    1400 4. Fem1      1400    1446     1423          1
#> 4_10   f    1489 4. Fem1      1489    1529     1509          1
#> 4_11   f    1500 4. Fem1      1500    1539     1520          1
#> 4_12   f    1510 4. Fem1      1510    1548     1529          1
#> 4_13   f    1521 4. Fem1      1521    1558     1540          1
#> 4_14   f    1530 4. Fem1      1530    1566     1548          1
#> 4_15   f    1539 4. Fem1      1539    1575     1557          1
#> 4_16   f    1551 4. Fem1      1551    1585     1568          1
#> 4_17   f    1560 4. Fem1      1560    1594     1577          1
#> 4_18   f    1569 4. Fem1      1569    1602     1586          1
#> 4_19   f    1580 4. Fem1      1580    1613     1597          1
#> 4_2    f    1409 4. Fem1      1409    1455     1432          1
#> 4_20   f    1589 4. Fem1      1589    1621     1605          1
#> 4_21   f    1601 4. Fem1      1601    1632     1616          1
#> 4_22   f    1610 4. Fem1      1610    1640     1625          1
#> 4_23   f    1619 4. Fem1      1619    1649     1634          1
#> 4_24   f    1630 4. Fem1      1630    1659     1645          1
#> 4_25   f    1640 4. Fem1      1640    1668     1654          1
#> 4_26   f    1651 4. Fem1      1651    1678     1665          1
#> 4_27   f    1660 4. Fem1      1660    1687     1673          1
#> 4_28   f    1669 4. Fem1      1669    1695     1682          1
#> 4_29   f    1681 4. Fem1      1681    1706     1693          1
#> 4_3    f    1421 4. Fem1      1421    1465     1443          1
#> 4_30   f    1690 4. Fem1      1690    1714     1702          1
#> 4_31   f    1701 4. Fem1      1701    1725     1713          1
#> 4_32   f    1710 4. Fem1      1710    1733     1722          1
#> 4_33   f    1719 4. Fem1      1719    1742     1730          1
#> 4_34   f    1731 4. Fem1      1731    1752     1741          1
#> 4_35   f    1740 4. Fem1      1740    1761     1750          1
#> 4_36   f    1749 4. Fem1      1749    1769     1759          1
#> 4_37   f    1760 4. Fem1      1760    1780     1770          1
#> 4_38   f    1770 4. Fem1      1770    1788     1779          1
#> 4_39   f    1781 4. Fem1      1781    1799     1790          1
#> 4_4    f    1430 4. Fem1      1430    1474     1452          1
#> 4_40   f    1790 4. Fem1      1790    1807     1799          1
#> 4_41   f    1799 4. Fem1      1799    1815     1807          1
#> 4_42   f    1811 4. Fem1      1811    1826     1818          1
#> 4_43   f    1820 4. Fem1      1820    1834     1827          1
#> 4_44   f    1831 4. Fem1      1831    1845     1838          1
#> 4_45   f    1840 4. Fem1      1840    1853     1847          1
#> 4_5    f    1439 4. Fem1      1439    1482     1461          1
#> 4_6    f    1450 4. Fem1      1450    1493     1471          1
#> 4_7    f    1459 4. Fem1      1459    1501     1480          1
#> 4_8    f    1471 4. Fem1      1471    1512     1491          1
#> 4_9    f    1480 4. Fem1      1480    1520     1500          1
#>